+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 0.3.2 (unreleased)
 
 * Use `google-api-python-client` 2.7.0
-* `File#download_file` and `File#download_workbook` now return `None` instead of `False` if `._client` isn’t set
+* `File#download_file`, `File#download_workbook` and `File#get_child` now return `None` instead of `False`
+  if `._client` isn’t set
+* `File#exists` now returns `None` instead of `False` if `._client` isn’t set
 * `File#download_workbook` now has a correct return type hint: `Optional[Workbook]` instead of `bool`
 * Remove three `client.Client` private methods (`_serialize_query_value`, `_make_query_clause`, `_make_querystring`)